fix: review of User,Event,Partipate,Child

+++ enpoints/enpoints/src/main/java/com/parentlink/service/EventService.java
@@ -4,7 +4,10 @@ import com.parentlink.model.Event;
 import com.parentlink.repository.EventRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -26,14 +29,56 @@ public class EventService {
     }
 
     public Event updateEvent(Long id, Event eventDetails) {
-        Event event = eventRepository.findById(id).orElseThrow(() -> new RuntimeException("Event not found"));
-        event.setName(eventDetails.getName());
-        event.setLocation(eventDetails.getLocation());
-        event.setDate(eventDetails.getDate());
+        // Buscar el evento en la base de datos por su ID
+        Event event = eventRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("Event not found with id: " + id));
+
+        // Actualizar solo los campos que no sean null en eventDetails
+        if (eventDetails.getName() != null) {
+            event.setName(eventDetails.getName());
+        }
+
+        if (eventDetails.getDescription() != null) {
+            event.setDescription(eventDetails.getDescription());
+        }
+
+        if (eventDetails.getImage() != null) {
+            event.setImage(eventDetails.getImage());
+        }
+
+        if (eventDetails.getMinAge() != null) {
+            event.setMinAge(eventDetails.getMinAge());
+        }
+
+        if (eventDetails.getMaxAge() != null) {
+            event.setMaxAge(eventDetails.getMaxAge());
+        }
+
+        if (eventDetails.getDate() != null) {
+            event.setDate(eventDetails.getDate());
+        }
+
+        if (eventDetails.getLocation() != null) {
+            event.setLocation(eventDetails.getLocation());
+        }
+
+        // Guardar el evento con los cambios aplicados
         return eventRepository.save(event);
     }
 
+    // Eliminar un evento por su ID
     public void deleteEvent(Long id) {
+        // Usamos NoSuchElementException para verificar la existencia del evento
+        if (!eventRepository.existsById(id)) {
+            throw new NoSuchElementException("Event not found with id: " + id);
+        }
         eventRepository.deleteById(id);
+    }
+
+    public boolean canAddRemark(Event event) {
+        LocalDateTime currentDateTime = LocalDateTime.now();
+        LocalDateTime eventEndDate = event.getDate().plusDays(1); // Sumar un día al evento
+
+        return currentDateTime.isAfter(eventEndDate); // Si la fecha actual es, al menos, un día después del evento
     }
 }
